feat: auth flow
